Fixes for form and functionality
To Drayton: having fixed links for CSS is bad, because when we move our
project to the server, all these links will be broken(our URL will not
be localhost anymore). Please be careful when dealing with absolute
links.

Also: put css into header template, and JS into footer template. Please
do not put CSS for the page in the content section.